+++ tests/test_glib.py
@@ -1,0 +1,27 @@
+import unittest
+import glib
+
+
+class TestGlib(unittest.TestCase):
+
+    def test_check_directory(self):
+        pass
+
+    def test_create_folder(self):
+        pass
+
+    def test_list_subdirectories(self):
+        pass
+
+    def test_list_files(self):
+        pass
+
+    def test_list_shortcuts(self):
+        pass
+
+    def test_print_shortcuts(self):
+        pass
+
+
+if __name__ == '__main__':
+    unittest.main()