+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.3.0] - TBA
 # Changed
+- Added notifications for added or removed textblocks
 - Reorganized classes so that methods make more sense
 - Renamed internal use methods according to PEP-8
 - Renamed Settings.py as ConfigUtils.py as this is more accurate